build: zero warnings across every preset and both compilers
255 warnings on a full gcc build (701 on clang) down to 0 on all six presets —
gcc/clang x debug/dev/ship — with tests green on each.

Most were not ours, and are silenced at the source rather than by weakening the
warning set:
- ImGui's vendored backends are compiled into Assisi-Debug, so they inherited
  Assisi::Warnings and produced ~55 conversion warnings in upstream code (55).
- reflectgen emits FieldMeta initializers that stop at the last member a field
  needs, leaving the rest to their default member initializers; correct, but
  flagged once per omitted member (125).
- doctest's TEST_CASE expands __COUNTER__, doctest::Approx takes a double so
  every float comparison promotes, and it includes <ciso646> which libstdc++ 16
  #warns about — ~650 warnings under clang, none from our logic. Test targets
  now use Assisi_apply_test_defaults().
- glslang's own sources warn under GCC at -O2/-O3 (12, ship only).

An ordering trap worth recording: -Wno- flags set with target_compile_options
land BEFORE the ones Assisi::Warnings supplies as INTERFACE options, so the
later -Wextra silently re-enabled them. Clang did this; gcc did not, which is
why the generated-code suppression appeared to work at first. Every suppression
here is applied per-source, which is appended last.

Ours, actually fixed:
- ~30 sign-conversion warnings from indexing arrays with an int that the
  surrounding API (ImGui/ImPlot Offset and count, GLFW key codes) requires to
  stay signed. Converted at the index, after the existing bounds check, rather
  than retyping variables whose type is dictated by that API.
- float->double promotions made explicit at printf-family and ImPlot call
  sites; kLoadingFps became double since its only use multiplies GetTime().
- AssetSidecar::Leaf(id) replaces AssetSidecar{.guid = id} at 9 sites, naming
  the concept the struct's own docs describe (empty/absent for a leaf).
- A %llu format string that my earlier explicit-width sweep had invalidated:
  the fields became uint64_t, which is unsigned long on Linux. Now PRIu64.
- A dead file-local kMaxMaterials alias (shadowed by the class member inside
  AssetCache's own methods), an unneeded constexpr lambda capture, a duplicate
  Clock alias, and an unused object-layer count.

One suppression is a false positive, scoped to a single file: GCC reports a
maybe-uninitialized read inside nlohmann's move constructor when a
ComponentDelta (two optional<json>) moves into the command variant, in
optimized builds only. Every member is explicitly initialized and restructuring
the call does not move it.

Verified: ASSISI_WARNINGS_AS_ERRORS=ON now builds clean, so the option is real
rather than aspirational. Left OFF by default — flipping it is your call.

+++ apps/sandbox/src/SandboxAssetBrowser.cpp
@@ -163,7 +163,8 @@ void DrawMaterialIcon(const ImVec2 &origin, float size)
 // sequence, or a Private-Use codepoint like 0xE000).
 constexpr uint32_t     kTtfFirstFrame = 0xF000; // Spinner.ttf frames: U+F000..U+F12B
 constexpr int32_t      kTtfFrameCount = 300;    // one full v4-30 atom-spin loop (must match the font)
-constexpr float        kLoadingFps    = 30.0f;  // 300 frames / 30fps = 10 s per loop (matches the design).
+constexpr double       kLoadingFps    = 30.0;   // 300 frames / 30fps = 10 s per loop (matches the design).
+                                                // double: its only use multiplies ImGui::GetTime(), which is double.
                                                 // NOTE: this drives BOTH backends -- the WebP path ignores
                                                 // the durations baked into the file -- so it must match
                                                 // whichever v4 variant is shipped as the asset.
